requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# PDF processing
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0

# Computer vision and layout detection
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
ultralytics==8.0.206

# OCR
pytesseract==0.3.10
transformers==4.35.2
datasets==2.14.6

# NLP and embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2

# Database and search
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pgvector==0.2.4
faiss-cpu==1.7.4
whoosh==2.7.4

# Utilities
numpy==1.24.4
pandas==2.1.3
tqdm==4.66.1
python-dotenv==1.0.0
click==8.1.7
loguru==0.7.2
pyyaml==6.0.1
jinja2==3.1.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: GPU acceleration
# torch==2.1.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision==0.16.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
